moved db model to it's own file; moved config settings to __init__.py
