+++ app.py
@@ -1,31 +1,34 @@
-from flask_api import FlaskAPI, status
+# from flask_api import FlaskAPI, status
+# import dummyapi.token as token
+# from flask_sqlalchemy import SQLAlchemy
+# import os
+
+# app = FlaskAPI(__name__)
+# base_dir = os.path.abspath(os.path.dirname(__file__))
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///'\
+#      + os.path.join(base_dir, 'data.sqlite')
+# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+# db = SQLAlchemy(app)
+
+# class ApiKey(db.Model):
+#     key = db.Column(db.String(255), primary_key=True)
+
+#     def __init__(self, key):
+#         self.key = key
+
+#     def json(self):
+#         return {'key': self.key}
+
+
+from dummyapi import app, db
+from dummyapi.models import ApiKey
 import dummyapi.token as token
-from flask_sqlalchemy import SQLAlchemy
-import os
-
-app = FlaskAPI(__name__)
-base_dir = os.path.abspath(os.path.dirname(__file__))
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///'\
-     + os.path.join(base_dir, 'data.sqlite')
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy(app)
-
-
+from flask_api import FlaskAPI, status
 # I'd like to keep deployment as-is 
-# with this function users do not need to build the db from CLI
+# # with this function users do not need to build the db from CLI
 @app.before_first_request
 def create_tables():
     db.create_all()
-
-
-class ApiKey(db.Model):
-    key = db.Column(db.String(255), primary_key=True)
-
-    def __init__(self, key):
-        self.key = key
-
-    def json(self):
-        return {'key': self.key}
 
 
 @app.route("/token", methods=["GET"])
